inner sorted join
